Hide Achievement creation behind a helper
As the Achievement struct grows, I don't want to have to adjust each
Rule implementation. Since a Rule is really just deciding whether an
Achievement should be granted to a Commit, and the contents of the
Achievements won't vary based on the Rules (but it's overridable if they
do), using a trait method with a default impl seems best.