Add test for retaining state between repositories
I think it's too error prone to require each Rule implementation to
remember to reset its internal state on calls to `finalize()`. It would
be better to re-create new impl Rule objects for each repository.

This is what crate::achievement::process_rules::grant() does. But it's a
bit inflexible (especially when I get to customizing the rules that are
run, and injecting user-contrib rules, and opinionated ones).

But then this gets into how crate::achievements::builtin_rules() works,
which I want to change up to use something more akin to a plugin
registry (no centralized list, rules can register themselves) using the
inventory crate.

But I ran into lots of problems with that, because Rule implementations
need to be mutable so they can store state, but also they need to be
Sized trait objects (so Box<dyn Rule>) except that inventory requires
const fns, which Box::new() isn't.

So I think the plan is to use inventory to register and collect Rule
factories, rather than the Rules themselves. But that's pretty tricky,
and needs to change up how the main application iteration works.

This will also impact how parallelism works, when I get to it.