Merge test_foo.rs -> foo.rs
I tried splitting out the tests so that changes in the tests wouldn't
trigger recompilation of the tested module, but that was an
overoptimization.

I seem to recall making this choice from matklad's excellent One Hundred
Thousand Lines of Rust [1] series, but I can't find it again, and I've
come to not liking having them split into different modules.

[1] https://matklad.github.io/2021/09/05/Rust100k.html