Reproduce fetch bug in a failing test
There were several mutations attempted that all relate to this bug:
* changing += to *=
* changing == to !=
* etc
all in the region that calculates the number of new commits.

It turns out that counting the number of new commits only works the
first time the branch is pulled, and that subsequent updates don't
actually update the local branch. This means that the desired workflow
of

    herostratus add ...
    herostratus fetch-all
    herostratus check-all
    ... time goes by
    herostratus fetch-all
    herostratus check-all

doesn't actually work as-expected.

This might be a tricky bug to fix, and I might end up choosing to not
fix it until I finish migrating to gix. The value is in having the
tests, since fetching has turned up being more complex than imagined.

TODO: There should be an integration test covering the desired workflow.