Refactor Achievements::get_next_achievement_online
Before, it would only process one Rule at a time for each commit. But
this is painful for diff processing, because it makes it hard to share
a single diff across all rules.

So refactor it to process all rules at once for each commit, because
then it'll be easier to compute a single diff and share it across all
rules.